# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentación de mi API Rest\n\nPara autenticarte:\n1. Usa el endpoint /auth/login para obtener el token\n2. Copia el token devuelto\n3. Click en el botón 'Authorize' (🔓) arriba\n4. Pega el token en el campo 'Value' (incluye 'Bearer ')",
        "title": "Viking-App ApiREST",
        "contact": {
            "name": "mirazopablo",
            "email": "mirazopablo@gmail.com"
        },
        "version": "1.0"
    },
    "host": "0.0.0.0:8080",
    "basePath": "/",
    "paths": {
        "/api/": {
            "get": {
                "security": [
                    {
                        "bearer-jwt": []
                    }
                ],
                "description": "Mensaje de bienvenida",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Home Controller"
                ],
                "summary": "Saludo",
                "operationId": "greeting",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/budget/by-work-order/{workOrderId}": {
            "get": {
                "security": [
                    {
                        "bearer-jwt": []
                    }
                ],
                "description": "Obtiene la estructura del presupuesto para su renderizado Web en tiempo real",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget"
                ],
                "summary": "Obtener presupuesto por ID de Orden de Trabajo",
                "operationId": "getBudgetByWorkOrder",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Work Order UUID",
                        "name": "workOrderId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Is Public Request",
                        "name": "public",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BudgetResponseDto"
                        }
                    }
                }
            }
        },
        "/api/budget/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "bearer-jwt": []
                    }
                ],
                "description": "Elimina físicamente un presupuesto de la base de datos por su UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget"
                ],
                "summary": "Eliminar presupuesto (Hard Delete)",
                "operationId": "deleteBudget",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Budget UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/budget/save": {
            "post": {
                "security": [
                    {
                        "bearer-jwt": []
                    }
                ],
                "description": "Almacena o actualiza la estructura JSON de un presupuesto y registra auditoría en diagnostic_points",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget"
                ],
                "summary": "Guardar o actualizar presupuesto",
                "operationId": "saveBudget",
                "parameters": [
                    {
                        "description": "Budget payload",
                        "name": "budget",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BudgetSaveDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BudgetResponseDto"
                        }
                    }
                }
            }
        },
        "/api/budget/update-status/{id}": {
            "patch": {
                "security": [
                    {
                        "bearer-jwt": []
                    }
                ],
                "description": "Permite actualizar el estado del presupuesto (DRAFT, SENT, APPROVED, REJECTED)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Budget"
                ],
                "summary": "Actualizar estado del presupuesto",
                "operationId": "updateBudgetStatus",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Budget UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status payload",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BudgetStatusUpdateDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BudgetResponseDto"
                        }
                    }
                }
            }
        },
        "/api/device/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "bearer-jwt": []
                    }
                ],
                "description": "Elimina un dispositivo de la base de datos",
                "tags": [
                    "Device Controller"
                ],
                "summary": "Eliminar Dispositivo",
                "operationId": "deleteDevice",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Device UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/device/save": {
            "post": {
                "security": [
                    {
                        "bearer-jwt": []
                    }
                ],
                "description": "Guarda un dispositivo en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device Controller"
                ],
                "summary": "Guardar Dispositivos",
                "operationId": "registerDevice",
                "parameters": [
                    {
                        "description": "Device Create Request",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DeviceCreateRequestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DeviceResponseDto"
                        }
                    }
                }
            }
        },
        "/api/device/search": {
            "get": {
                "security": [
                    {
                        "bearer-jwt": []
                    }
                ],
                "description": "Busca dispositivos en la base de datos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device Controller"
                ],
                "summary": "Buscar Dispositivos",
                "operationId": "searchDevice",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Device UUID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Serial Number",
                        "name": "serialNumber",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Brand",
                        "name": "brand",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "User DNI (búsqueda parcial incremental)",
                        "name": "userDni",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "User Name (búsqueda parcial incremental)",
                        "name": "userName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Selector de campo o término general",
                        "name": "query",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DeviceResponseDto"
                            }
                        }
                    }
                }
            }
        },
        "/api/device/update/{id}": {
            "patch": {
                "security": [
                    {
                        "bearer-jwt": []
                    }
                ],
                "description": "Actualiza un dispositivo existente de forma parcial o total",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device Controller"
                ],
                "summary": "Actualizar Dispositivo (Soporta PATCH parcial para reasignación de dueño)",
                "operationId": "updateDevice",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Device UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Device Update Request",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DeviceUpdateRequestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DeviceResponseDto"
                        }
                    }
                }
            }
        },
        "/api/diagnostic-points/add": {
            "post": {
                "security": [
                    {
                        "bearer-jwt": []
                    }
                ],
                "description": "Agrega un nuevo punto de diagnóstico a una orden de trabajo",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DiagnosticPoint"
                ],
                "summary": "Agregar punto de diagnóstico",
                "operationId": "addDiagnosticPoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Work Order UUID",
                        "name": "workOrderId",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Client UUID",
                        "name": "clientId",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Diagnostic description or JSON string",
                        "name": "diagnosticPoint",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Attached image file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DiagnosticPointResponseDto"
                        }
                    }
                }
            }
        },
        "/api/diagnostic-points/by-work-order/{workOrderId}": {
            "get": {
                "security": [
                    {
                        "bearer-jwt": []
                    }
                ],
                "description": "Obtiene una lista de puntos de diagnóstico por ID de orden de trabajo sin acoplamiento a clientId",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DiagnosticPoint"
                ],
                "summary": "Obtener puntos de diagnóstico por ID de orden de trabajo",
                "operationId": "getDiagnosticPointsByWorkOrder",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Work Order UUID",
                        "name": "workOrderId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DiagnosticPointResponseDto"
                            }
                        }
                    }
                }
            }
        },
        "/api/diagnostic-points/by-work-order/{workOrderId}/client/{clientId}": {
            "get": {
                "security": [
                    {
                        "bearer-jwt": []
                    }
                ],
                "description": "Obtiene una lista de puntos de diagnóstico por ID de orden de trabajo y Cliente Asociado a la orden de trabajo",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DiagnosticPoint"
                ],
                "summary": "Obtener puntos de diagnóstico por ID de orden de trabajo",
                "operationId": "getDiagnosticPointsByWorkOrderAndClient",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Work Order UUID",
                        "name": "workOrderId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Client UUID",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DiagnosticPointResponseDto"
                            }
                        }
                    }
                }
            }
        },
        "/api/diagnostic-points/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "bearer-jwt": []
                    }
                ],
                "description": "Borra una evidencia multimedia por UUID",
                "tags": [
                    "DiagnosticPoint"
                ],
                "summary": "Eliminar Punto de Diagnóstico",
                "operationId": "deleteDiagnosticPoint",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Diagnostic Point UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/user-roles/is-staff": {
            "get": {
                "security": [
                    {
                        "bearer-jwt": []
                    }
                ],
                "description": "Verifica si el usuario es staff",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Role Controller"
                ],
                "summary": "Verificar si el usuario es staff",
                "operationId": "isUserStaff",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
        "/api/user-roles/user-permission": {
            "get": {
                "security": [
                    {
                        "bearer-jwt": []
                    }
                ],
                "description": "Verifica si el usuario es admin",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Role Controller"
                ],
                "summary": "Verificar si el usuario es admin",
                "operationId": "getUserPermission",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/autocomplete": {
            "get": {
                "security": [
                    {
                        "bearer-jwt": []
                    }
                ],
                "description": "Devuelve una proyección ligera de usuarios para selectores sin exponer PII sensible",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Controller"
                ],
                "summary": "Autocompletado Rápido de Usuario / Cliente",
                "operationId": "autocompleteUser",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Término de búsqueda (nombre, DNI o teléfono)",
                        "name": "query",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserAutocompleteDto"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.BudgetResponseDto": {
            "type": "object",
            "properties": {
                "budgetData": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mode": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "totals": {
                    "$ref": "#/definitions/models.BudgetTotalsDto"
                },
                "updatedAt": {
                    "type": "string"
                },
                "workOrderId": {
                    "type": "string"
                }
            }
        },
...
        "models.BudgetStatusUpdateDto": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "bearer-jwt": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}